requirements.txt
fastapi
uvicorn
groq
requests
python-dotenv
python-multipart
numpy
